add two sound files client + server done, i have to do script execution functions inside and dlloading

+++ R-Type/shared/includes/Commands/CommandDestroyResource.hpp
@@ -9,7 +9,7 @@ class CommandDestroyResource : public NoCopyable, public ICommand {
 	public:
 		#pragma pack(push, 1)
 		struct PacketFromServer {
-			int32_t	id;
+			uint64_t id;
 		};
 		#pragma pack(pop)
 
@@ -20,8 +20,8 @@ class CommandDestroyResource : public NoCopyable, public ICommand {
 
 	// getter - setters
 	public:
-		int 	getId(void) const;
-		void	setId(int id);
+		uint64_t 	getId(void) const;
+		void		setId(uint64_t id);
 
 	// public pure methods
 	public:
@@ -32,6 +32,6 @@ class CommandDestroyResource : public NoCopyable, public ICommand {
 
 	// attributes
 	private:
-		int mId;
+		uint64_t mId;
 
 };
