+++ R-Type/shared/includes/Commands/CommandShowLevel.hpp
@@ -12,7 +12,7 @@ class CommandShowLevel : public NoCopyable, public ICommand {
 		#pragma pack(push, 1)
 		struct PacketFromServer {
 			char		name[64];
-			int32_t		scriptSize;
+			uint64_t	scriptSize;
 			char		script[4096];
 		};
 		#pragma pack(pop)
