+++ R-Type/shared/includes/Commands/CommandTimeElapsedPing.hpp
@@ -9,7 +9,7 @@ class CommandTimeElapsedPing : public NoCopyable, public ICommand {
 	public:
 		#pragma pack(push, 1)
 		struct PacketFromServer {
-			int64_t	timeElapsed;
+			double timeElapsed;
 		};
 		#pragma pack(pop)
 
@@ -20,8 +20,8 @@ class CommandTimeElapsedPing : public NoCopyable, public ICommand {
 
 	// getter - setters
 	public:
-		int64_t getTimeElapsed(void) const;
-		void		setTimeElapsed(int64_t timeElapsed);
+		double		getTimeElapsed(void) const;
+		void		setTimeElapsed(double timeElapsed);
 
 	// public pure methods
 	public:
@@ -32,6 +32,6 @@ class CommandTimeElapsedPing : public NoCopyable, public ICommand {
 
 	// attributes
 	private:
-		int64_t mTimeElapsed;
+		double mTimeElapsed;
 
 };
