+++ R-Type/shared/sources/Commands/SharedCommandMoveResource.cpp
@@ -1,12 +1,12 @@
 #include "CommandMoveResource.hpp"
 
-CommandMoveResource::CommandMoveResource(void) : mId(0), mType(IResource::Type::UNKNOWN), mX(0), mY(0), mAngle(0) {
+CommandMoveResource::CommandMoveResource(void) : mId(0), mType(IResource::Type::UNKNOWN), mX(0.0), mY(0.0), mAngle(0) {
 }
 
 CommandMoveResource::~CommandMoveResource(void) {
 }
 
-int CommandMoveResource::getId(void) const {
+uint64_t CommandMoveResource::getId(void) const {
 	return mId;
 }
 
@@ -14,11 +14,11 @@ IResource::Type CommandMoveResource::getType(void) const {
 	return mType;
 }
 
-float CommandMoveResource::getX(void) const {
+double CommandMoveResource::getX(void) const {
 	return mX;
 }
 
-float CommandMoveResource::getY(void) const {
+double CommandMoveResource::getY(void) const {
 	return mY;
 }
 
@@ -26,7 +26,7 @@ short CommandMoveResource::getAngle(void) const {
 	return mAngle;
 }
 
-void	CommandMoveResource::setId(int id) {
+void	CommandMoveResource::setId(uint64_t id) {
 	mId = id;
 }
 
@@ -34,11 +34,11 @@ void	CommandMoveResource::setType(IResource::Type type) {
 	mType = type;
 }
 
-void	CommandMoveResource::setX(float x) {
+void	CommandMoveResource::setX(double x) {
 	mX = x;
 }
 
-void	CommandMoveResource::setY(float y) {
+void	CommandMoveResource::setY(double y) {
 	mY = y;
 }
